notebooks/fraudfinder/scripts/cloud_run_model_inference/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-api-python-client>=1.7.8,<2
google-cloud-aiplatform
google-cloud-pubsub
pandas